+++ cloud-sql/mysql/client-side-encryption/src/test/java/cloudsql/tink/QueryDecryptDataIT.java
@@ -67,11 +67,12 @@ public class QueryDecryptDataIT {
   public static void setUp() throws GeneralSecurityException, SQLException {
     checkEnvVars();
     tableName = String.format("votes_%s", UUID.randomUUID().toString().replace("-", ""));
-    pool = QueryAndDecryptData
-        .createConnectionPool(MYSQL_USER, MYSQL_PASS, MYSQL_DB, MYSQL_CONNECTION_NAME);
-    QueryAndDecryptData.createTable(pool, tableName);
 
-    envAead = new CloudKmsEnvelopeAead(CLOUD_KMS_URI).envAead;
+    pool = CloudSqlConnectionPool
+        .createConnectionPool(MYSQL_USER, MYSQL_PASS, MYSQL_DB, MYSQL_CONNECTION_NAME);
+    CloudSqlConnectionPool.createTable(pool, tableName);
+
+    envAead = CloudKmsEnvelopeAead.getEnvelopeAead(CLOUD_KMS_URI);
     EncryptAndInsertData
         .encryptAndInsertData(pool, envAead, tableName, "TABS", "hello@example.com");
   }
